created reservation route and remove api from main.py
